Process: Add missing type casts.
Keep the compiler happy to avoid Travis build failing.

+++ src/process.c
@@ -734,7 +734,7 @@ cc_oci_vm_launch (struct cc_oci_config *config)
 			goto child_failed;
 		}
 
-		hypervisor_args = g_new0(gchar, 1+hypervisor_args_len);
+		hypervisor_args = g_new0(gchar, (gsize)(1+hypervisor_args_len));
 		if (! hypervisor_args) {
 			g_critical ("failed alloc memory for hypervisor args");
 			goto child_failed;
@@ -879,7 +879,7 @@ child_failed:
 		goto out;
 	}
 
-	hypervisor_args_len = g_utf8_strlen(hypervisor_args, -1);
+	hypervisor_args_len = (gint)g_utf8_strlen(hypervisor_args, -1);
 
 	/* first - write hypervisor length */
 	bytes = write (hypervisor_args_pipe[1], &hypervisor_args_len,
